lxd/operations: Disconnect from DB on non-LXD
